Extract ToNXevent_data into file
