+++ src/beamlime/handlers/accumulators.py
@@ -2,7 +2,6 @@
 # Copyright (c) 2025 Scipp contributors (https://github.com/scipp)
 from __future__ import annotations
 
-from collections.abc import Sequence
 from dataclasses import dataclass
 from typing import Any, TypeVar
 
@@ -10,51 +9,11 @@ import numpy as np
 import scipp as sc
 from ess.reduce.live.roi import ROIFilter
 from scipp._scipp.core import _bins_no_validate
-from streaming_data_types import eventdata_ev44, logdata_f144
+from streaming_data_types import logdata_f144
 
 from ..config import models
 from ..core.handler import Accumulator, Config, ConfigModelAccessor
-
-
-@dataclass
-class MonitorEvents:
-    """
-    Dataclass for monitor events.
-
-    Decouples our handlers from upstream schema changes. This also simplifies handler
-    testing since tests do not have to construct a full eventdata_ev44.EventData object.
-
-    Note that we keep the raw array of time of arrivals, and the unit. This is to avoid
-    unnecessary copies of the data.
-    """
-
-    time_of_arrival: Sequence[int]
-    unit: str
-
-    @staticmethod
-    def from_ev44(ev44: eventdata_ev44.EventData) -> MonitorEvents:
-        return MonitorEvents(time_of_arrival=ev44.time_of_flight, unit='ns')
-
-
-@dataclass
-class DetectorEvents(MonitorEvents):
-    """
-    Dataclass for detector events.
-
-    Decouples our handlers from upstream schema changes. This also simplifies handler
-    testing since tests do not have to construct a full eventdata_ev44.EventData object.
-
-    Note that we keep the raw array of time of arrivals, and the unit. This is to avoid
-    unnecessary copies of the data.
-    """
-
-    pixel_id: Sequence[int]
-
-    @staticmethod
-    def from_ev44(ev44: eventdata_ev44.EventData) -> DetectorEvents:
-        return DetectorEvents(
-            pixel_id=ev44.pixel_id, time_of_arrival=ev44.time_of_flight, unit='ns'
-        )
+from .to_nxevent_data import DetectorEvents, MonitorEvents
 
 
 @dataclass
@@ -72,70 +31,6 @@ class LogData:
     @staticmethod
     def from_f144(f144: logdata_f144.LogData) -> LogData:
         return LogData(time=f144.timestamp_unix_ns, value=f144.value)
-
-
-Events = TypeVar('Events', DetectorEvents, MonitorEvents)
-
-
-class ToNXevent_data(Accumulator[Events, sc.DataArray]):
-    def __init__(self):
-        self._chunks: list[Events] = []
-        self._timestamps: list[int] = []
-        self._epoch = sc.epoch(unit='ns')
-        self._have_event_id: bool | None = None
-        self._toa_dtype = np.int64
-
-    def add(self, timestamp: int, data: Events) -> None:
-        if data.unit != 'ns':
-            raise ValueError(f"Expected unit 'ns', got '{data.unit}'")
-        if self._have_event_id is None:
-            self._have_event_id = isinstance(data, DetectorEvents)
-            self._toa_dtype = np.asarray(data.time_of_arrival).dtype
-        elif self._have_event_id != isinstance(data, DetectorEvents):
-            # This should never happen, but we check to be safe.
-            raise ValueError("Inconsistent event_id")
-        if isinstance(data, DetectorEvents) and (
-            len(data.time_of_arrival) != len(data.pixel_id)
-        ):
-            raise ValueError("Inconsistent data length")
-        self._chunks.append(data)
-        self._timestamps.append(timestamp)
-
-    def get(self) -> sc.DataArray:
-        if self._have_event_id is None:
-            raise ValueError("No data has been added")
-        if self._chunks:
-            toa_values = np.concatenate([d.time_of_arrival for d in self._chunks])
-        else:
-            toa_values = np.array([], dtype=self._toa_dtype)
-        event_time_offset = sc.array(dims=['event'], values=toa_values, unit='ns')
-        weights = sc.ones(sizes=event_time_offset.sizes, dtype='float32', unit='counts')
-        events = sc.DataArray(
-            data=weights, coords={'event_time_offset': event_time_offset}
-        )
-        if self._have_event_id:
-            if self._chunks:
-                ids = np.concatenate([d.pixel_id for d in self._chunks])
-            else:
-                ids = np.array([])
-            event_id = sc.array(dims=['event'], values=ids, unit=None, dtype='int32')
-            events.coords['event_id'] = event_id
-
-        lens = [len(d.time_of_arrival) for d in self._chunks]
-        sizes = sc.array(
-            dims=['event_time_zero'], values=lens, unit=None, dtype='int64'
-        )
-        begin = sc.cumsum(sizes, mode='exclusive')
-        binned = sc.DataArray(sc.bins(begin=begin, dim='event', data=events))
-        binned.coords['event_time_zero'] = self._epoch + sc.array(
-            dims=['event_time_zero'], values=self._timestamps, unit='ns', dtype='int64'
-        )
-        self.clear()
-        return binned
-
-    def clear(self) -> None:
-        self._chunks.clear()
-        self._timestamps.clear()
 
 
 class NullAccumulator(Accumulator[Any, None]):
